move: layout 컴포넌트 타입 합치기
